Add files via upload
Challenge:
Try typecasting the return value from dequeueReusableAnnotationView() so that it's an MKPinAnnotationView. Once that’s done, change the pinTintColor property to your favorite UIColor.
Add a UIAlertController that lets users specify how they want to view the map. There's a mapType property that draws the maps in different ways. For example, .satellite gives a satellite view of the terrain.
Modify the callout button so that pressing it shows a new view controller with a web view, taking users to the Wikipedia entry for that city.